pyproject.toml
[project]
name = "algorithm-visualizer"
version = "0.1.0"
description = "A high-performance algorithm visualizer"
authors = [
    {name = "inconceivable"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "pygame-ce (>=2.5.6,<3.0.0)",
    "numpy (>=2.4.2,<3.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "scipy (>=1.17.0,<2.0.0)",
    "pygame-gui (>=0.6.14,<0.7.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
